requirements.txt
Django
djangorestframework
drf-spectacular
django-cors-headers
djangorestframework_simplejwt
djoser
python-dotenv
qdrant-client
sentence-transformers
numpy
pandas
tqdm
qdrant-client[fastembed]